shared_mem_cadence:
Figured out why the shared mem seemingly was not getting read from on the
cadence side of things. The library was compiled for a 32 bit arch so the sizes
of the structs were different from the 64 bit driver process. After compiling
the driver process as a 32 bit executable, the communication worked from driver
-> sim.