shared_mem_cadence: Automation Improvements
Checkpointing work, but am automating the flow to launch the sim and interact
with the sim. The sim is now launched through docker in a python module. The
module also provides wrappers to the functions in interprocess.cpp that can be
called from python with pybind11. Need to complete the vsim module and use in
gem5.