package_impedence: Package Model
Modeled the package impedence of the processor according to the intel EVSD 11.1
(2009). The motherboard and the socket have an impedence between the output of
the VRM and the processor die. This helps to smooth out the slew rate of the
processors current draw seen at the supply. This also has python scripts to
plot fft, cdf and time domains of the output signals.