circuit_model: Compiling ARM PDN.
This commit brings updates to the README, as well as entirely restructures the
circuit model directory. Common modules are now in common/ and power
distribution network models are now in pdn/. All of the warings from NCVerilog
are cleaned up by storing a return value of adding compiler guards to the top
of the files. This code compiles but is untested.