harvard_pdn, arm_pdn: Finalized
The ARM PDN impedence profile now matches the 2015 paper, and the harvard PDN
is now comparable to the DeCoR papers power supply impedence profile.

Both have been verified in Cadence Spectre AC Sim and Transient Sim, QUCS AC
Sim and Transient Sim.

Furthermore, there is a utility for designing the harvard PDN LC values, and
now the verilog-A code uses proper transition filter analog operator and
a current source to model the load.

Finally an additional option was added to the simulation control file,
reducing the ringing on the nets during transient simulation.