full-system:
Added a few new scripts, zipping up the image files, and uploading them to
drive. It seems there is a bug in the restore from checkpoint and switch CPU to
DerivO3CPU where multiple interripts from the APIC cause the simulation to
either throw and exeption for bad_function_call, or after implementing a hash
of the interrupt and its corresponding function call, it hangs.

Switching over to SE Mode with Ruby and Cooperative Multithreadding.